+++ custom_components/wallbox_gateway/__init__.py
@@ -31,6 +31,7 @@ PLATFORMS: list[Platform] = [
     Platform.NUMBER,
     Platform.SELECT,
     Platform.BUTTON,
+    Platform.UPDATE,
 ]